# factor out f32<->f16 conversion to C library