also added fast non-APFloat implementations for both f16 and bf16

verified exact match with APFloat with exhaustive test, invoked with `--exhaustive` flag

put the implementations in a C library to prepare for adding f16 and bf16 support into cruntime (not in this PR)

moving from a macro to a function made f32->f16 conversion 1.5% slower on my MacBook M1 which uses clang's bulitin `__fp16` conversion, and made f32->bf16 and bf16->f32 conversion 19x and 8x faster

before:
```
$ RelWithDebInfo/bin/TestSmallFP
Benchmark                          Time             CPU   Iterations
--------------------------------------------------------------------
BM_F32_TO_FP16<float_16>      138918 ns       138897 ns         4818
BM_F32_TO_FP16<bfloat_16>    2658403 ns      2658372 ns          258
BM_FP16_TO_F32<float_16>      224751 ns       224743 ns         3127
BM_FP16_TO_F32<bfloat_16>    1826754 ns      1826755 ns          383
```

after:
```
Benchmark                          Time             CPU   Iterations
--------------------------------------------------------------------
BM_F32_TO_FP16<float_16>      140891 ns       140885 ns         4770
BM_F32_TO_FP16<bfloat_16>     139581 ns       139581 ns         5000
BM_FP16_TO_F32<float_16>      224199 ns       224199 ns         3120
BM_FP16_TO_F32<bfloat_16>     225306 ns       225306 ns         3121
```

generic f32->f16 and f16->f32 is 16x and 8x faster on my MacBook M1 when I disable clang's bulitin `__fp16` conversion

before:
```
BM_F32_TO_FP16<float_16>     2369281 ns      2369282 ns          291
BM_FP16_TO_F32<float_16>     1823704 ns      1823703 ns          384
```

after:
```
BM_F32_TO_FP16<float_16>      140401 ns       140386 ns         4673
BM_FP16_TO_F32<float_16>      224145 ns       224110 ns         3113
```