@AlexandreEichenberger 2023-07-07
> Moving forward it might make sense to add some some benchmarks. We have Google bench as a third party dependency.

look in the test/perf subdirectories for examples