@sorenlassen 2023-07-07
> > Moving forward it might make sense to add some some benchmarks. We have Google bench as a third party dependency.
> 
> look in the test/perf subdirectories for examples

yes, TestSmallFP has google benchmark tests which produced the performance numbers quoted in the PR description